Include artifacts differing only by classifier or and/scope (#145)
When folding through configurations and filtering out already included
modules, only module id was considered. That meant that if we had the
same module id for example in 'compile' and 'test' configurations, with
different classifiers, the one on test configuration was dropped out
as it was already 'seen' in compile configuration.

Now we filter on artifact level instead of module level, so that
artifact classifier is taken into account.

This patch fixes also a duplicate sources -issue, if you used both
withSources() and gen-idea with classifiers.